second commit ready for posting

+++ admin/includes/view_all_users.php
@@ -83,7 +83,7 @@
 
 if (isset($_GET['change_to_admin'])) {
 
-	$the_user_id = $_GET['change_to_admin'];
+	$the_user_id = escape($_GET['change_to_admin']);
 
 	$query = "UPDATE users SET user_role = 'admin' WHERE user_id = $the_user_id";
 	$admin_user_query = mysqli_query($connection, $query);
@@ -94,7 +94,7 @@ if (isset($_GET['change_to_admin'])) {
 
 if (isset($_GET['change_to_subscriber'])) {
 
-	$the_user_id = $_GET['change_to_subscriber'];
+	$the_user_id = escape($_GET['change_to_subscriber']);
 
 	$query = "UPDATE users SET user_role = 'subscriber' WHERE user_id = $the_user_id";
 	$subscriber_user_query = mysqli_query($connection, $query);
@@ -105,11 +105,16 @@ if (isset($_GET['change_to_subscriber'])) {
 
 if (isset($_GET['delete'])) {
 
-	$the_user_id = $_GET['delete'];
+	if (isset($_SESSION['user_role'])) {
 
-	$query = "DELETE FROM users WHERE user_id = {$the_user_id}";
-	$delete_query = mysqli_query($connection, $query);
-	header('Location: users.php');
+		if ($_SESSION['user_role'] == 'admin') {
+			$the_user_id = mysqli_real_escape_string($connection, escape($_GET['delete']));
 
-	confirm($delete_query);
+			$query = "DELETE FROM users WHERE user_id = {$the_user_id}";
+			$delete_query = mysqli_query($connection, $query);
+			header('Location: users.php');
+
+			confirm($delete_query);
+		}
+	}
 }
